refactor tests: separate endpoint test in different files
